+++ examples/optimizers/run_bwoa_example.py
@@ -5,7 +5,7 @@
 # --------------------------------------------------%
 
 import numpy as np
-from mealpy import FloatVar, BWOA, APO, GRSA, KLA, MGOA, AAA, NWOA, OSA, DandelionO, RFO, CrayfishOA
+from mealpy import FloatVar, BWOA, APO, GRSA, KLA, MGOA, AAA, NWOA, OSA, DandelionO, RFO, CrayfishOA, SPBO, CCO
 
 
 def objective_function(solution):
@@ -32,6 +32,8 @@ model = DandelionO.OriginalDandelionO(epoch=1000, pop_size=50)
 model = DandelionO.DevDandelionO(epoch=1000, pop_size=50)
 model = RFO.OriginalRFO(epoch=1000, pop_size=50, phi_0=0.7, theta=0.5)
 model = CrayfishOA.OriginalCrayfishOA(epoch=1000, pop_size=50)
+model = SPBO.OriginalSPBO(epoch=1000, pop_size=50)
+model = CCO.OriginalCCO(epoch=1000, pop_size=50, alpha=0.5, beta=1.0)
 
 g_best = model.solve(problem, seed=10)
 print(f"Best fitness: {g_best.target.fitness}")